requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.3
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
kombu==5.3.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
